First commit. ver 1.0
The BOT only does full test in order of the official SAT. All breaks are included. 
Will add the ability to pick and choose sections in the future. 

Use command ‘!SAT’ to start a test 
Command ‘!STOP’ to stop the bot at any time 

Command ‘!DEMO’ is used to demo the bot and Times are in Seconds rather than mins (i.e 6.5 mins for the Reading section instead of 65 mins)